% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{bolt_connect}
\alias{bolt_connect}
\title{Open a connection. \code{uri} is a Bolt URI, e.g. "neo4j://localhost:7687"
or "127.0.0.1:7687". When \code{database} is non-empty it selects the default
database for the session; an empty string uses the server default
("neo4j"). Returns an external pointer that R holds onto.}
\usage{
bolt_connect(uri, user, password, database)
}
\description{
Open a connection. \code{uri} is a Bolt URI, e.g. "neo4j://localhost:7687"
or "127.0.0.1:7687". When \code{database} is non-empty it selects the default
database for the session; an empty string uses the server default
("neo4j"). Returns an external pointer that R holds onto.
}
