\name{scoreCrunchBase}
\alias{scoreCrunchBase}
\title{Score persons}
\usage{
  scoreCrunchBase(count = 10000)
}
\value{
  a data.frame with columns "person", "coworker" and
  "score"
}
\description{
  Score is based on the total money raised of a company,
  discounted according to person's job title.
}

