\name{scoreSchool}
\alias{scoreSchool}
\title{Add a numeric value feature to the data frame}
\usage{
  scoreSchool(edu, scores = NULL)
}
\arguments{
  \item{edu}{input dataset with a column named "school"}
}
\value{
  dataset with added column called "school_score"
}
\description{
  Adds a new column based on the names of school
}

