\name{weightScoreByTitle}
\alias{weightScoreByTitle}
\title{Weight score by factor based on title}
\usage{
  weightScoreByTitle(results)
}
\arguments{
  \item{results}{dataframe must have "score" and "title"
  columns}
}
\value{
  data.frame with scores adjusted
}
\description{
  Weight score by factor based on title
}

